Add support for formatting source from stdin (`uaua fmt -`)
Add a check to format_file to treat the filename `-` as standard
input. If this is provided, it must also necessarily write to
standard output, so a `-` implies `--to-stdout` for the file.

This is useful for editors like Helix that expect a formatter to
work entirely over stdin and stdout.